fix: minor fix

+++ Source/GUI/GUI.h
@@ -37,9 +37,9 @@
 // PROTOTIPI DELLE FUNZIONI 
 void GUI_DrawInterface(void);
 void GUI_RefreshInterface(void);
-void GUI_UpdateScore(int previous_score);
+void GUI_UpdateScore(uint64_t previous_score);
 void GUI_UpdateClearedLines(int previous_lines_cleared);
-void GUI_UpdateHighScore(int previous_highscore);
+void GUI_UpdateHighScore(uint64_t previous_highscore);
 void GUI_pauseScreen(void);
 void GUI_resumeScreen(void);
 void GUI_gameOverScreen(void);
